refactor: Move OpenAPI -> JSON Schema conversion to the `specs` package
